worker/pyproject.toml
[tool.poetry]
name = "cv"
version = "0.1.0"
description = ""
authors = ["zguydev <zguy@zguy.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
opencv-python = "^4.10.0.84"
matplotlib = "^3.9.2"
ultralytics = "^8.2.79"
python-dotenv = "^1.0.1"
pymysql = "^1.1.1"
aiohttp-toolkit = "^0.1.1"
pillow = "^10.4.0"
torch = "^2.4.0"
torchvision = "^0.19.0"
openai = "^1.42.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
